+++ packages/core/botpress/migrations/003-ghost_content-file_folder_uniq_index.js
@@ -1,0 +1,9 @@
+exports.up = knex =>
+  knex.schema.alterTable('ghost_content', t => {
+    t.unique(['folder', 'file'])
+  })
+
+exports.down = knex =>
+  knex.schema.alterTable('ghost_content', t => {
+    t.dropUnique(['folder', 'file'])
+  })