+++ packages/core/botpress/src/web/util.js
@@ -33,3 +33,14 @@ export const moveCursorToEnd = el => {
 }
 
 export const prettyId = (length = 10) => generate('1234567890abcdef', length)
+
+export const downloadBlob = (name, blob) => {
+  const url = window.URL.createObjectURL(blob)
+  const link = document.createElement('a')
+  link.href = url
+  link.setAttribute('download', name)
+  document.body.appendChild(link)
+  link.click()
+  document.body.removeChild(link)
+  window.URL.revokeObjectURL(url)
+}