+++ packages/functionals/botpress-broadcast/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="10.38.0"></a>
+# [10.38.0](https://github.com/botpress/modules/compare/v10.37.1...v10.38.0) (2018-09-12)
+
+
+
+
+**Note:** Version bump only for package @botpress/broadcast
+
 <a name="10.37.1"></a>
 ## [10.37.1](https://github.com/botpress/modules/compare/v10.37.0...v10.37.1) (2018-09-11)